+++ restart_endpoint_worker.js
@@ -19,49 +19,42 @@ if (!port || port < 3000) {
 }
 
 cp.execSync('./stopcontainer.sh ' + port, (error, stdout, stderr) => {
-    // catch err, stdout, stderr
     if (error) {
         console.log('Error in removing files');
         return;
     }
     if (stderr) {
-        console.log('has stderr output');
+        console.log('Has stderr output');
         console.log(stderr);
-        // return;
     }
 
     return;
 });
 
 try {
-    fs.unlinkSync('runningports.txt');
-    //file removed
+    if (fs.existsSync('runningports.txt')) {
+        fs.unlinkSync('runningports.txt');
+    }
 } catch (err) {
     console.error(err);
 }
 
 cp.exec('./removestopped.sh', (error, stdout, stderr) => {
-    // catch err, stdout, stderr
     if (error) {
         console.log('Error in removing files');
-        // return;
     }
     if (stderr) {
-        console.log('has stderr output');
+        console.log('Has stderr output');
         console.log(stderr);
-        // return;
     }
 
     cp.exec('./runstopped.sh', async (error, stdout, stderr) => {
-        // catch err, stdout, stderr
         if (error) {
             console.log('Error in removing files');
-            // return;
         }
         if (stderr) {
-            console.log('has stderr output');
+            console.log('Has stderr output');
             console.log(stderr);
-            // return;
         }
 
         const runningSet = new Set();
@@ -90,19 +83,15 @@ cp.exec('./removestopped.sh', (error, stdout, stderr) => {
                     if (!runningSet.has(toRun)) {
                         cp.exec('./restartstopped.sh ' + portRow.service_name + portRow.service_endpoint + ':1.0 ' + toRun
                             + ' ' + portRow.service_name, (error, stdout, stderr) => {
-                                // catch err, stdout, stderr
                                 if (error) {
                                     console.log('Error in removing files');
-                                    // return;
                                 }
                                 if (stderr) {
-                                    console.log('has stderr output');
+                                    console.log('Has stderr output');
                                     console.log(stderr);
-                                    // return;
                                 }
 
-                                // endpointReadyMap.set(portRow.service_endpoint, true);
-                                parentPort.postMessage({endpointSegment: portRow.service_endpoint, ready: true});
+                                parentPort.postMessage({ endpointSegment: portRow.service_endpoint, ready: true });
                             });
                     }
